translator_backend/requirements.txt
# ══════════════════════════════════════════════
# NMT — Fase 2 | requirements.txt completo
# ══════════════════════════════════════════════

# ── EXISTENTES (Fase 1) ──────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
groq>=0.9.0
edge-tts>=6.1.9
miniaudio>=1.59
python-dotenv>=1.0.0
websockets>=12.0

# ── BANCO DE DADOS ───────────────────────────
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# ── AUTENTICAÇÃO ─────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# ── HTTP CLIENT (AbacatePay) ─────────────────
httpx>=0.27.0

# ── RATE LIMITING ────────────────────────────
slowapi>=0.1.9

# ── EMAIL ────────────────────────────────────
resend>=2.0.0
